Notes for the previous commit:
In the end, I have decided to use BuckleScript for compiling from
OCaml to JavaScript. This was because the JavaScript FFI on js_of_ocaml
is exceptionally poorly documented, and the build system is fairly
confusing.

Another decision I have made is to write the entire project in OCaml.
This is because it is really cumbersome to pass comlex objects between
JavaScript and OCaml, whichever binding I use, because I would have to
convert between types representable in JavaScrip and types
representable in OCaml. Additionally, examples for this do exist.

Switching to BuckleScript had the added advantage of very fast compiles,
and as a corollary, a "watch" mode on the build system, which compiles
each file as they are changed. I am exploiting this in my makefile.

A disadvantage of BuckleScript is that BuckleScript compiles each file
into a separate module. This works well for for example Node.js, but
most browsers don't have a module system built in. Therefore, one has
to either package all modules into a single JavaScript file with a tool
like webpack, or somehow get a module system into the browser.

Webpack works reasonably well, and it does also have a watch mode.
However, I it is not really convenient to debug all the code in a
single monolithic file. (Another shortcoming of BuckleScript is that it
doesn't produce source maps. However, it does produce some fairly
readable code.) BuckleSript can produce ECMAScript 6 modules, which do
work in Firefox when used with the appropriate configuration option.
However, the JavaScript files BS produces use module imports, which are
not handled by Firefox, as the ES6 module system is not fully defined
yet.

Another option is using Require.js and AMD modules, which can also be
generated by BuckleScript. I have not tried this.