Added a generated module with the revision number of the commit being built.
This will be useful when I am trying out the interface and collecting inputs.
Each session will be tagged with the exact revision number, which will ensure
that I can reproduce the exact session from the inputs.

The generating script runs before every build. Additionally, it is ran by
Chokidar after any source file changes, to ensure that the dirty flag is
applied. Additionally, it is ran as a git post-commit hook to apply the new
revision number if a commit is done while a watch is running.