Finally make some sort of interactive demo
This demo has very little to do with any UI plans, but rather serves as a
proof that we can integrate the current backend with a HTML+Jquery interface
using Bucklescript.

The biggest hinderance here was actually getting bucklescript-jquery to work,
which involved me forking that library and submitting a pull request.