Actually log things
This is why I bothered so much with the revision numbers

I have also had to resolve two circular dependencies.
This is an interesting topic to talk about in a thesis.
Also, if this happens again:

https://wiki.xen.org/wiki/OCaml_Cyclical_Build_Dependencies

Another bit of mess in the last commit was the bad naiming
of json-codec, which broke module resolution and
the lack of support for qualified unions in the same.