Add array formation operator
This has to be an n-ary operator, and this induced a big wave of
rewrites. However, if I went for a binary operator, I would have had to
deal with extending the array on the left one by one.

Additionally some tests and resulting bugfixes that really don't belong
in this commit but I can't be bothered.