Recursive functions
I have opted for making recursive functions separate form lets, because
I found it easier to implement it this way, and because I don't want
the recursive definition of any other data type to be possible.

This however results in the recursive variable being potentially
different to the variable that it is being assigned to (if at all
assigned). While this is a theoretically correct view, it might be
surprising for some.